use `startProcess` instead of `execCmdEx`, add `shellVerbose`
When using the macro in non NimScript mode, it now uses `startProcess`
to start a proper child process. This allows us to read from that
processs' stdout and echo it in real time. Handy for long running
processes!

Also `shellVerbose` is added, which will return full output of the
commands. That is a tuple of the total echoed output and the exit
code.
Note that the exit code will either be 0 or refer to the last not
failed exit code. If any command fails, the subsequent commands will
not be run!