quote via () instead of ``
Quoting via `()` allows for easier concatenation of Nim identifiers
and string literals, which for many practical use cases (at least in
my experience) was pretty lacking with the previous accent quote based
approach.

Usage of () in bash is limited to subshell execution, which seems a
reasonable sacrifice. If something complicated enough is to be done
with this macro to warrant the bash usage of (), using a literal
string for this should be fine.