use `repr` instead of `strVal`
At this point in the code we expect that the argument given here is
actually what the user would normally put into the `&{}` string
interpolation macro. This means we actually just want the string
representation of that `NimNode`, because that way we support more
complex statements, e.g. handing a proc call, tuple field access etc.

If the user hands random rubbish this will at the latest fail in the
string interpolation macro anyways.