change `shellVerbose` impl, allow processOptions to be set
This is quite a big change to the internals of `shellVerbose`. Essentially it
reduces the number of internal procedures used to a single `shellVerboseImpl`
with a pre parsing stage of the possible arguments (see below) in `shellVerbose`.

In addition one can now customize the process options, which are handed to
`startProcess`.

Finally, the process is now using `poEvalCommand` by default.

`shellVerbose` parses macro arguments from `args: varargs[untyped]`, where it
expects to get up to 3 parameters in addition to an `nnkStmtList`.

The arguments are debugConfig, processOptions and combineOutAndErr. See the
docstring for details.